Exclude `Rule#ObjectFilter` from db column mapping
`Rule#ObjectFilter` is of type `*object.Filter` and doesn't have any
defined tags and is set its field name to `object_filter` by default.
`Rule#ObjectFilterExpr` on ther other hand defines a tag that is named
the same as the first one (`object_filter`) and thus the second field
isn't taken into account in `DB#BuildColumns()` since it's a duplicate
of the first field, which results in `Rule#ObjectFilter` not being loaded
from the database at all.