eventstream: Host/Service Events from CheckResult
Turns out, the timestamp from Event Stream StateChange objects is
neither reflected in the CheckResult nor in a Host or Service's
last_state_change field. I'm not quite sure where a CheckResult's
timestamp originates, but now I switched to rely on the CheckResult
which is identical in both the Event Stream as well as in Hosts and
Services.

While doing so, the duplicate Event generation logic was unified and
refactored into an own method.