eventstream: generic Attrs for ObjectQueriesResult
First, this change brings the advantage that the JSON unmarshalling does
not require any special code and works directly without double parsing
the input. Also, no double type check is needed later in the code,
because the type information is already there.

However, to make this work, some refactoring was necessary all over the
client_api.go place. A major obstacle was that methods are not allowed
to have type parameters. Thus, some logic needed to be extracted out of
the Client type into an own function. This resulted in changes for the
HTTP calling methods and a separation of HTTP transmission and parsing
the response spreading throughout the library code.