eventstream: current time for new Events
Adapt the Event generation to the Listener's and use time.Now() for the
Event.Time. Furthermore, cleanup two Event generating methods for the
Event Stream which are only used once and have, after prior refactoring,
an empty function body.

Note: After this change, the caching is de facto broken, but will be
addressed in the next commit.