eventstream: reconnection in Event Stream method
First, the reconnection phase was renamed to the replay phase as it
handles events to be replayed, also during the initial connection.

The waitForApiAvailability method, which fired short lived HTTP probes
to check the API's availability, was removed. This logic was now
integrated into the listenEventStream method, which retries connecting
until a connection can be established. However, to not have to wait too
long for timeouts, the net.Dialer.Timeout needed to be adjusted.

Furthermore, as listenEventStream was the only remaining method in the
client_es.go file, it was moved to client_api.go, which now contains all
API related methods.