eventstream: rework replay communication logic
Prior, two variables - one signaling channel and an atomic.Bool - were
holding and/or communicating if the eventDispatcher was in its replay
phase. Especially the atomic.Bool variable was accessed both in the
producer - enterReplayPhase - as well as the consumer - eventDispatcher.

After this rework, the whole logic went into the main worker, previously
named eventDispatcher. By using a single channel - replayPhaseRequest -,
the worker might now switch to the replay phase. This eases and unifies
the internal "API" as all communication with the worker takes place over
unidirectional channels.

Within the worker, an internal channel is used to communicate replayed
events back from the producing goroutines to the worker. Also, when all
work is done or a guarding context is closed, this channel is closed,
which itself is a signal to switch modes again. This context is also
used to cancel already running replay jobs if another one was requested.