Drop `incident#AddHistory()` method
It was already quite annoying having to return `types.Int` every single
time this method was used. This commit embeds the functionality for
`insert and fetch ID` directly into the `HistoryRow` type and nowhere
needs to be returned. The just inserted and fetched ID can be accessed
with `history.ID`.