Query timeperiod_entry directly into the final struct
This adds the required fields and tags to timeperiod.Entry so that this struct
can be used directly for querying the database. This removes the previous
inline struct TimeperiodEntry that was present in the config loading code. All
the initialization is now done in the Init() method instead of when copying the
values between the structs.

This is done in preparation for changes to schedules where timeperiod entries
are loaded directly as part of schedules.