Define schedules using rotations
This is a fundamental and incompatible change to how schedules are defined.
Now, a schedule consists of a list of rotations that's ordered by priority.
Each rotation contains multiple members where each is either a contact or a
contact group. Each member is linked to some timeperiod entries which defines
when this member is active in the rotation.

This commit already includes code for a feature that was planned but is
possible using the web interface at the moment: multiple versions of the same
rotation where the handoff time defines when a given version becomes active.

With this change, for the time being, the TimePeriod type itself fulfills no
real purpose and the timeperiod entries are directly loaded as part of the
schedule, bypassing the timeperiod loading code. However, there still is the
plan to add standalone timeperiods in the future, thus the timeperiod code is
kept.

More context for these changes:
- https://github.com/Icinga/icinga-notifications-web/issues/177
- https://github.com/Icinga/icinga-notifications/pull/193