Simplify schema upgrade for schedule rotations
Reduce the schema file to the minimal changes that are needed to go from the
old to the new version. This makes it easier to follow what has been changed
and also allows keeping other timeperiods (if they were manually created as it
was not yet possible to create those outside of schedules using the web
interface so far). Comments were removed from here as they still exist in the
full schema file and the purpose of the upgrade file is not to document the
schema.