schema.sql, 025.sql: make style consistent
Mainly adapt spacing and character case to be consistent with the rest. Also do
the CREATE EXTENSION at the top of the schema file, it may be used elsewhere in
the future as well. One functional change: rename the primary key constraint to
be consistent with pk_${table_name} used in the rest of the file.