Use `ctx` everywhere & render exit errors properly
The `context.Context` related changes are neccessary as otherwise you
will have to wait for 5m when having Icinga notification started with
invalid database config due to the internal database.Connect retrials.
Lastly, the logging related changes are useful as we want to properly
render the stack traces when we unexpectedly exit from the main function
instead of wrapping the error in a confusing `zap.Error()` json key.