icinga2: Check errors in HTTP status and JSON
The Icinga 2 Event Stream connection did not validate the HTTP status
code returned by the server. Consequently, HTTP errors were not
identified and their error messages were incorrectly interpreted as
valid messages. This occurred in two places: firstly, with regard to the
HTTP status code, and secondly, in relation to the error field, which
may be present within a JSON struct.

This modification introduces a brief HTTP status validation method that
adheres to the specifications outlined in the Icinga 2 API documentation
and advanced error checks in UnmarshalEventStreamResponse.