Require manually calling Init() for timeperiod entries
Lazily initializing the timeperiod entries comes with problems regarding the
error handling and would also result in race conditions if multiple callers use
the object at the same time. This commit changes this so that Init() has to be
called explicitly first, allowing proper error handling and read-only use of
the object later.