Distinct "internal" Event Types
It was ambiguous what the "internal" event type especially stand for. As
currently both of its use cases are some kind of reevaluation - either
during the program's startup or due to an event's age -, it was renamed
to "incident-age".

To ensure database-level stability, the column type was changed from
text to a custom enum.

Closes #162.