Add Makefile
It's a very basic Makefile providing just the all/test/install/clean targets
intended to avoid duplication of the build and installation steps in the
Dockerfile and the rpm/deb packaging repos.

/usr/sbin was chosen for the installation path because that's where the Icinga
DB packages place the binary.

By using the Makefile, the Docker image will now also include the default
config and the schema files.