Add default path for config file
Previously, the path to the config file had to be passed on the command line.
This commit adds a sensible default so that this is no longer necessary.

As the mechanism for setting the channel plugin dir based on libexecdir from
the Makefile is already there, the same was used for sysconfdir, even though
that should always be /etc for our packaging needs.