Don't set `server#WriteTimeout`
Setting `10s` as a write timeout might not cause any problems when the
request is processed successfully, but when we e.g. have to retry any
database errors, which is `5m` by default, this will just in an i/o
timeout and the listener won't event write any hints about that, but will
silently abort the connections.