channels: Reorder methods according to plugin.Plugin
The plugin.Plugin interface specifies the three methods required for the
channel plugin RPC API in their logical order. Considering that these
channels are the blueprint for external channel development, the order
of implementing channel plugins was changed to correspond to the
interface.

For all files, the main function has been moved to the top.

In addition, the Email.Send method received a documentation string
indicating that it implements the enmime.Sender interface, as it is not
directly called anywhere. Further, the Email.GetServer method got
inlined as it was only called from one other method.