icinga2: Common Muted Event for Acknowledgements w/o Comments during Catch-Up
When creating an acknowledgement in Icinga Web, a comment is also added.
It is possible, however, to later delete the comment while keeping the
object acknowledged.

Deleting acknowledgement comments was not an expected behavior,
requiring the presence of an associated comment for each
acknowledgement. Otherwise, the catch-up-phase would not succeed.

Unfortunately, there is no way to later add an author to an
acknowledgement unless a matching comment exists. Thus, the
catch-up-phase will now generalize this acknowledgement into a common
muted event, missing the details available in the comment.

Closes #245.