internal/object: Fix gosec G115
The conversion of an int64 to an uint64 was marked as a potential
overflow, since it might lose precision. However, in this case this
should be of no issue since only the binary representation matters, not
the actual numeric value.

So a "nolint" tag was added next to an explanatory comment.

Fixes #289.