Channel: Recreate Logger After Name Change
When restarting a channel, recreate its logger based on the current
fields since they may have changed due to reconfiguration.

As a related change, use the Channel.MarshalLogObject method instead of
populating the fields in incrementalApplyPending's createFn handler.
Unfortunately, zap caches the result from the zap.Object function call
and not re-evaluates it every time the logger is used. That is the
reason for recreating the logger in the Channel.Restart method.

Fixes #309.