config: NoRulesVersion for initial comparisons
Change the rule version string for empty or no rules from the wayward
"N/A" string to an empty string.

Doing so lets comparisons with an empty rule version evaluate to true,
which conveniently reduces the amount of rule exchanges between a source
and this daemon on a clean setup.

In addition, explicitly set the version to NoRulesVersion. This was also
done in Listener.writeSourceRulesInfo, even if this now has no real
effect. However, before this method returned an empty string version,
while the rule comparison expected "N/A", resulting in rejected
submissions.