Listener: add /incidents endpoint
Add an API endpoint for sources to query all incidents belonging to them. This
allows sources to detect incidents for obsolete objects, i.e. objects that
disappeared or were deleted while they had an ongoing incident.

Added primarily for Icinga for Kubernetes, but could be useful for other
sources as well.